refactor: move claims walking code to published package
